refactor: 리팩토링 수정 사항 진행 (#27)
