+++ src/main/java/com/hid_web/be/storage/exhibit/ExhibitRepository.java
@@ -1,4 +1,4 @@
-package com.hid_web.be.storage;
+package com.hid_web.be.storage.exhibit;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
